Make impersonate_user local to a course
If a user is a super-user, they'll be able to impersonate anyone.
If not, a user (ta) be able to impersonate a user who they have access
to (student) in that particular course.

Note that when impersonating, a ta will only be able to access the
courses of that student that he/she has permission for. Access to
other courses will be denied. (Any urls starting with /course/<id>)

One possible issue with this approach is that a ta will be able to
see the home page of the student, which means they'll be able to see
the titles of some private courses that the student has access to.